Add utils for float casting ergonomics

+++ crates/cargo-truce/src/scaffold/kind.rs
@@ -127,7 +127,7 @@ pub struct {struct_name}Params {
 const EFFECT_PROCESS_BODY: &str = r"    fn process(&mut self, buffer: &mut AudioBuffer, _events: &EventList,
                _context: &mut ProcessContext) -> ProcessStatus {
         for i in 0..buffer.num_samples() {
-            let gain = db_to_linear(self.params.gain.smoothed_next() as f64) as f32;
+            let gain = db_to_linear(self.params.gain.smoothed_next());
             for ch in 0..buffer.channels() {
                 let (inp, out) = buffer.io(ch);
                 out[i] = inp[i] * gain;
